반복되는 유사한 코드를 하나의 컴포넌트로 통합
